refactor: ProfilePageからLearning GoalsセクションをProfileGoalsSectionに分離
ProfilePage: 381→344行（37行削減）
